move pdf into passes folder and update namespacing
